+++ test/libsolidity/smtCheckerTests/functions/getters/struct_4.sol
@@ -11,7 +11,7 @@ contract C {
 
 	function test() public view {
 		(D d, function () external returns (uint) f) = this.s();
-		assert(d == s.d); // should hold
+		assert(address(d) == address(s.d)); // should hold
 		assert(address(d) == address(this)); // should fail
 	}
 }
@@ -19,4 +19,4 @@ contract C {
 // SMTEngine: all
 // ----
 // Warning 2072: (146-183): Unused local variable.
-// Warning 6328: (234-269): CHC: Assertion violation happens here.\nCounterexample:\ns = {d: 0, f: 0}\nd = 0\nf = 0\n\nTransaction trace:\nC.constructor()\nState: s = {d: 0, f: 0}\nC.test()
+// Warning 6328: (252-287): CHC: Assertion violation happens here.\nCounterexample:\ns = {d: 0, f: 0}\nd = 0\nf = 0\n\nTransaction trace:\nC.constructor()\nState: s = {d: 0, f: 0}\nC.test()