+++ test/libsolidity/syntaxTests/conversion/implicit_conversion_of_super_in_comparison.sol
@@ -20,8 +20,8 @@ contract C {
 }
 // ----
 // TypeError 2271: (144-157): Operator != not compatible with types type(contract super C) and contract C
-// TypeError 2271: (167-180): Operator != not compatible with types contract C and type(contract super C)
+// TypeError 2271: (167-180): Operator != not compatible with types contract C and type(contract super C). No arithmetic or comparison operations are allowed on contract types. Consider converting to "address".
 // TypeError 2271: (254-264): Operator != not compatible with types type(contract super C) and contract C
-// TypeError 2271: (274-284): Operator != not compatible with types contract C and type(contract super C)
+// TypeError 2271: (274-284): Operator != not compatible with types contract C and type(contract super C). No arithmetic or comparison operations are allowed on contract types. Consider converting to "address".
 // TypeError 2271: (349-359): Operator != not compatible with types type(contract super C) and contract D
-// TypeError 2271: (369-379): Operator != not compatible with types contract D and type(contract super C)
+// TypeError 2271: (369-379): Operator != not compatible with types contract D and type(contract super C). No arithmetic or comparison operations are allowed on contract types. Consider converting to "address".